consider sick note extension with half work days
extending sick notes with a half working day length is not possible
(while not catched currently...)

this is just to be defensive I think

+++ src/main/java/org/synyx/urlaubsverwaltung/sicknote/sicknote/SickNoteExtendDto.java
@@ -1,6 +1,7 @@
 package org.synyx.urlaubsverwaltung.sicknote.sicknote;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
-record SickNoteExtendDto(Long id, LocalDate startDate, LocalDate endDate, long workingDays, boolean isAub) {
+record SickNoteExtendDto(Long id, LocalDate startDate, LocalDate endDate, BigDecimal workingDays, boolean isAub) {
 }
